Fix missing prototype in previous commit.

+++ dataobj/dingliste.cc
@@ -613,7 +613,7 @@ ding_t *dingliste_t::remove_last()
 	return d;
 }
 
-inline bool dingliste_t::_remove_by_index(unsigned i)
+inline bool dingliste_t::_remove_by_index(uint8 i)
 {
 	if (i >= top) {
 		// out of range
